Moved to separate .py file, fix docs
